requirements.txt
playwright==1.52.0
PyYAML==6.0.2
pydantic==2.11.3
rich==14.0.0
python-dotenv==1.1.0
pytest==9.0.2
